Update test_event_detection.py
Incorporating Ryan's feedback from code review. Note - I think cleaning up the section extracting the event latencies is going to require reconfiguring how we're saving the Matlab structure because the Python dictionaries are single-subject and the MATLAB structure-to-dictionary has all of them, so they can't be neatly unpacked one-to-one with a parallel loop/list comprehension approach.